Fix build breaks and sort out volatile
* fixed: build break caused by different prototype in header vs. source
* fixed: build break caused by typo
* changed: added volatile back into radio.c as in this case it is
  probably needed

+++ include/libopencm3/nrf/common/gpio.h
@@ -223,7 +223,7 @@ void gpio_set_options(uint32_t gpioport, uint32_t drive, uint32_t sense,
 		uint32_t gpios);
 
 void gpio_configure_task(uint8_t task_num,
-		uint8_t pin_num, uint8_t polarity, uint8_t init);
+		uint8_t pin_num, uint8_t polarity, uint32_t init);
 
 void gpio_configure_event(uint8_t event_num, uint8_t pin_num, uint8_t polarity);
